+++ generators/csharp/codegen/src/AsIs.ts
@@ -1,5 +1,6 @@
 export const STRING_ENUM_SERIALIZER_CLASS_NAME = "StringEnumSerializer";
 export const ONE_OF_SERIALIZER_CLASS_NAME = "OneOfSerializer";
+export const COLLECTION_ITEM_SERIALIZER_CLASS_NAME = "CollectionItemSerializer";
 
 export enum AsIsFiles {
     EnumConverter = "EnumConverter.Template.cs",
@@ -11,5 +12,6 @@ export enum AsIsFiles {
     RawClient = "RawClient.Template.cs",
     CiYaml = "github-ci.yml",
     StringEnumSerializer = "StringEnumSerializer.cs",
-    OneOfSerializer = "OneOfSerializer.cs"
+    OneOfSerializer = "OneOfSerializer.cs",
+    CollectionItemSerializer = "CollectionItemSerializer.cs"
 }