+++ packages/cli/openapi-ir-sdk/src/sdk/api/resources/finalIr/types/OpenApiIntermediateRepresentation.ts
@@ -22,4 +22,5 @@ export interface OpenApiIntermediateRepresentation {
     nonRequestReferencedSchemas: Set<FernOpenapiIr.SchemaId>;
     securitySchemes: Record<FernOpenapiIr.SecuritySchemeId, FernOpenapiIr.SecurityScheme>;
     globalHeaders: FernOpenapiIr.GlobalHeader[] | undefined;
+    idempotencyHeaders: FernOpenapiIr.IdempotencyHeader[] | undefined;
 }