+++ seed/csharp-model/examples/src/SeedExamples/Types/Tree.cs
@@ -8,5 +8,5 @@ namespace SeedExamples;
 public class Tree
 {
     [JsonPropertyName("nodes")]
-    public List<Node>? Nodes { get; init; }
+    public IEnumerable<Node>? Nodes { get; init; }
 }