+++ seed/csharp-model/pagination/src/SeedPagination/Users/UserListContainer.cs
@@ -8,5 +8,5 @@ namespace SeedPagination;
 public class UserListContainer
 {
     [JsonPropertyName("users")]
-    public List<User> Users { get; init; }
+    public IEnumerable<User> Users { get; init; }
 }