+++ seed/csharp-sdk/trace/src/SeedTrace/Playlist/Types/UpdatePlaylistRequest.cs
@@ -13,5 +13,5 @@ public class UpdatePlaylistRequest
     /// The problems that make up the playlist.
     /// </summary>
     [JsonPropertyName("problems")]
-    public List<string> Problems { get; init; }
+    public IEnumerable<string> Problems { get; init; }
 }