+++ seed/csharp-sdk/trace/src/SeedTrace/V2/Problem/ProblemClient.cs
@@ -18,7 +18,7 @@ public class ProblemClient
     /// <summary>
     /// Returns lightweight versions of all problems
     /// </summary>
-    public async Task<List<LightweightProblemInfoV2>> GetLightweightProblemsAsync()
+    public async Task<IEnumerable<LightweightProblemInfoV2>> GetLightweightProblemsAsync()
     {
         var response = await _client.MakeRequestAsync(
             new RawClient.ApiRequest { Method = HttpMethod.Get, Path = "/lightweight-problem-info" }
@@ -26,7 +26,7 @@ public class ProblemClient
         string responseBody = await response.Raw.Content.ReadAsStringAsync();
         if (response.StatusCode >= 200 && response.StatusCode < 400)
         {
-            return JsonSerializer.Deserialize<List<LightweightProblemInfoV2>>(responseBody);
+            return JsonSerializer.Deserialize<IEnumerable<LightweightProblemInfoV2>>(responseBody);
         }
         throw new Exception(responseBody);
     }
@@ -34,7 +34,7 @@ public class ProblemClient
     /// <summary>
     /// Returns latest versions of all problems
     /// </summary>
-    public async Task<List<ProblemInfoV2>> GetProblemsAsync()
+    public async Task<IEnumerable<ProblemInfoV2>> GetProblemsAsync()
     {
         var response = await _client.MakeRequestAsync(
             new RawClient.ApiRequest { Method = HttpMethod.Get, Path = "/problem-info" }
@@ -42,7 +42,7 @@ public class ProblemClient
         string responseBody = await response.Raw.Content.ReadAsStringAsync();
         if (response.StatusCode >= 200 && response.StatusCode < 400)
         {
-            return JsonSerializer.Deserialize<List<ProblemInfoV2>>(responseBody);
+            return JsonSerializer.Deserialize<IEnumerable<ProblemInfoV2>>(responseBody);
         }
         throw new Exception(responseBody);
     }