+++ seed/csharp-sdk/trace/src/SeedTrace/V2/V3/Problem/Types/VoidFunctionSignatureThatTakesActualResult.cs
@@ -9,7 +9,7 @@ namespace SeedTrace.V2.V3;
 public class VoidFunctionSignatureThatTakesActualResult
 {
     [JsonPropertyName("parameters")]
-    public List<Parameter> Parameters { get; init; }
+    public IEnumerable<Parameter> Parameters { get; init; }
 
     [JsonPropertyName("actualResultType")]
     public VariableType ActualResultType { get; init; }