+++ seed/csharp-sdk/undiscriminated-unions/src/SeedUndiscriminatedUnions/Union/UnionClient.cs
@@ -16,9 +16,23 @@ public class UnionClient
     }
 
     public async Task<
-        OneOf<string, List<string>, int, List<int>, List<List<int>>, HashSet<string>>
+        OneOf<
+            string,
+            IEnumerable<string>,
+            int,
+            IEnumerable<int>,
+            IEnumerable<IEnumerable<int>>,
+            HashSet<string>
+        >
     > GetAsync(
-        OneOf<string, List<string>, int, List<int>, List<List<int>>, HashSet<string>> request
+        OneOf<
+            string,
+            IEnumerable<string>,
+            int,
+            IEnumerable<int>,
+            IEnumerable<IEnumerable<int>>,
+            HashSet<string>
+        > request
     )
     {
         var response = await _client.MakeRequestAsync(
@@ -33,7 +47,14 @@ public class UnionClient
         if (response.StatusCode >= 200 && response.StatusCode < 400)
         {
             return JsonSerializer.Deserialize<
-                OneOf<string, List<string>, int, List<int>, List<List<int>>, HashSet<string>>
+                OneOf<
+                    string,
+                    IEnumerable<string>,
+                    int,
+                    IEnumerable<int>,
+                    IEnumerable<IEnumerable<int>>,
+                    HashSet<string>
+                >
             >(responseBody);
         }
         throw new Exception(responseBody);