+++ seed/python-sdk/oauth-client-credentials-default/tests/conftest.py
@@ -12,6 +12,7 @@ def client() -> SeedOauthClientCredentialsDefault:
         client_id=os.getenv("ENV_CLIENT_ID", "client_id"),
         client_secret=os.getenv("ENV_CLIENT_SECRET", "client_secret"),
         base_url=os.getenv("TESTS_BASE_URL", "base_url"),
+        _token_getter_override=lambda: os.getenv("ENV_TOKEN", "token"),
     )
 
 
@@ -21,4 +22,5 @@ def async_client() -> AsyncSeedOauthClientCredentialsDefault:
         client_id=os.getenv("ENV_CLIENT_ID", "client_id"),
         client_secret=os.getenv("ENV_CLIENT_SECRET", "client_secret"),
         base_url=os.getenv("TESTS_BASE_URL", "base_url"),
+        _token_getter_override=lambda: os.getenv("ENV_TOKEN", "token"),
     )