+++ generators/csharp/sdk/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.21 - 2024-05-31]
+
+- Fix: Array and List fields are now generated as `IEnumerable<>`. Additionally, if the 
+  item type is a `OneOf`, then a new core class called `CollectionItemSerializer` is used 
+  to deserialize the values. 
+
 ## [0.0.20 - 2024-05-29]
 
 - Fix: Enum serializer is now added to enum declarations instead of enum references. This